tests + wording
